+++ kernel/Utilities/BasisFunctions3DH1ConformingPrism.h
@@ -61,7 +61,7 @@ namespace Utilities
     {
     public:
         BasisFunction3DEdgePrism_0(std::size_t node0, std::size_t node1, std::size_t polynomialOrder)
-                : edgePosition_((node0 / 3) * 2 - 1), node0_(node0 % 3), node1_(node1 % 3), polynomialOrder_(polynomialOrder)
+                : edgePosition_((static_cast<int>(node0) / 3) * 2 - 1), node0_(node0 % 3), node1_(node1 % 3), polynomialOrder_(polynomialOrder)
         {
             logger.assert(node0 < 6, "A triangular prism only has 6 nodes");
             logger.assert(node1 < 6, "A triangular prism only has 6 nodes");
